文件重构
把ui分为了不同的单独文件，便于维护。同时，更改了ui更新的方式。
